Updated eslint and removed unnecessary file
