+++ .eslintrc
@@ -1,2 +1,8 @@
----
-  extends: braintree/client
+{
+  "extends": "braintree/client",
+  "env": {
+    "browser": true,
+    "node": true,
+    "es6": true
+  }
+}
