--- validator/index.js
@@ -1,127 +1,0 @@
-'use strict';
-
-var types = require('./cardTypes.json');
-var clone = require('./helpers/clone');
-var findBestMatch = require('./helpers/findBestMatch');
-var findType = require('./helpers/findType');
-var getAllCardTypes = require('./helpers/getAllCardTypes');
-var getCardPosition = require('./helpers/getCardPosition');
-var isValidInputType = require('./helpers/isValidInputType');
-var loopOverCardPatterns = require('./helpers/loopOverCardPatterns');
-
-var testOrder;
-var customCards = {};
-
-var cardNames = {
-  VISA: 'visa',
-  MASTERCARD: 'mastercard',
-  AMERICAN_EXPRESS: 'american-express',
-  DINERS_CLUB: 'diners-club',
-  DISCOVER: 'discover',
-  JCB: 'jcb',
-  UNIONPAY: 'unionpay',
-  MAESTRO: 'maestro',
-  ELO: 'elo',
-  MIR: 'mir'
-};
-
-var ORIGINAL_TEST_ORDER = [
-  cardNames.VISA,
-  cardNames.MASTERCARD,
-  cardNames.AMERICAN_EXPRESS,
-  cardNames.DINERS_CLUB,
-  cardNames.DISCOVER,
-  cardNames.JCB,
-  cardNames.UNIONPAY,
-  cardNames.MAESTRO,
-  cardNames.ELO,
-  cardNames.MIR
-];
-
-testOrder = clone(ORIGINAL_TEST_ORDER);
-
-function creditCardType(cardNumber) {
-  var bestMatch;
-  var results = [];
-
-  if (!isValidInputType(cardNumber)) {
-    return [];
-  }
-
-  if (cardNumber.length === 0) {
-    return getAllCardTypes(testOrder, customCards);
-  }
-
-  testOrder.forEach(function (type) {
-    var cardConfiguration = findType(type, customCards);
-
-    loopOverCardPatterns(cardNumber, cardConfiguration, results);
-  });
-
-  bestMatch = findBestMatch(results);
-
-  if (bestMatch) {
-    return [bestMatch];
-  }
-
-  return results;
-}
-
-creditCardType.getTypeInfo = function (type) {
-  return clone(findType(type, customCards));
-};
-
-creditCardType.removeCard = function (name) {
-  var position = getCardPosition(name, false, testOrder);
-
-  testOrder.splice(position, 1);
-};
-
-creditCardType.addCard = function (config) {
-  var existingCardPosition = getCardPosition(config.type, true, testOrder);
-
-  customCards[config.type] = config;
-
-  if (existingCardPosition === -1) {
-    testOrder.push(config.type);
-  }
-};
-
-creditCardType.updateCard = function (cardType, updates) {
-  var clonedCard;
-  var originalObject = customCards[cardType] || types[cardType];
-
-  if (!originalObject) {
-    throw new Error('"' + cardType + '" is not a recognized type. Use `addCard` instead.');
-  }
-
-  if (updates.type && originalObject.type !== updates.type) {
-    throw new Error('Cannot overwrite type parameter.');
-  }
-
-  clonedCard = clone(originalObject, true);
-
-  Object.keys(clonedCard).forEach(function (key) {
-    if (updates[key]) {
-      clonedCard[key] = updates[key];
-    }
-  });
-
-  customCards[clonedCard.type] = clonedCard;
-};
-
-creditCardType.changeOrder = function (name, position) {
-  var currentPosition = getCardPosition(name, false, testOrder);
-
-  testOrder.splice(currentPosition, 1);
-  testOrder.splice(position, 0, name);
-};
-
-creditCardType.resetModifications = function () {
-  testOrder = clone(ORIGINAL_TEST_ORDER);
-  customCards = {};
-};
-
-creditCardType.types = cardNames;
-
-module.exports = creditCardType;
